requirements.txt
torch==1.12.1
torchvision~=0.13.0
numpy
tqdm
tensorboard
matplotlib
pandas

speechbrain~=0.5.12
datasets
torch_audiomentations
wandb
torchaudio~=0.12.1
pillow
gdown
Unidecode
inflect
hparams